feat: Separate app.tsx
-Created Homepage component for content analyzer
-Developed Landingpage to introduce tools and its features
